requirements.txt
wget
numpy
PyMuPDF
rank_bm25
sentence-transformers
flask
python-telegram-bot
python-dotenv
pymongo
nltk

azure-storage-blob
azure-functions